requirements-ai.txt
langchain
langchain-openai
langchain-chroma
pandas
python-dotenv
langchain-community
lark